Refactor: OfficeInfoServiceApi를 MapViewModel로 이동
